Changed the organizarion of the code structure and changed the naming and moved to api error to be a method on internal exception. Changed the organizarion of the code structure and changed the naming and moved to api error to be a method on internal exception.
(#86c5dkb5w)
