+++ norman_objects/shared/exceptions/api_exceptions/norman_api_exception.py
@@ -1,9 +1,9 @@
 from datetime import datetime, UTC
 
-from norman_objects.shared.errors.norman_error import NormanError
+from norman_objects.shared.exceptions.norman_exception import NormanException
 
 
-class NormanApiError(NormanError):
+class NormanApiException(NormanException):
 
     status_code: int = 500
     error_type: str = "api_error"
@@ -12,15 +12,13 @@ class NormanApiError(NormanError):
     def __init__(
         self,
         message: str,
-        context: dict = None,
         timestamp: datetime = None,
         suggestions: list[str] = None
     ):
 
-        NormanError.__init__(
+        NormanException.__init__(
             self,
             message=message,
-            context=context if context is not None else {},
             timestamp=timestamp if timestamp is not None else datetime.now(UTC)
         )
 
@@ -28,12 +26,10 @@ class NormanApiError(NormanError):
             self.suggestions = suggestions
 
     def to_dict(self) -> dict:
-
         return {
             "message": self.message,
             "status_code": self.status_code,
             "error_type": self.error_type,
-            "context": self.context,
             "suggestions": self.suggestions,
             "timestamp": self.timestamp.isoformat()
         }
