+++ norman_objects/shared/exceptions/api_exceptions/rate_limit_exception.py
@@ -1,7 +1,7 @@
-from norman_objects.shared.errors.api_errors.norman_api_error import NormanApiError
+from norman_objects.shared.exceptions.api_exceptions.norman_api_exception import NormanApiException
 
 
-class RateLimitError(NormanApiError):
+class RateLimitException(NormanApiException):
     status_code: int = 429
     error_type: str = "rate_limit"
     suggestions: list[str] = [
