+++ norman_objects/shared/exceptions/internal_exceptions/database_exception.py
@@ -1,0 +1,5 @@
+from norman_objects.shared.exceptions.internal_exceptions.infrastructure_exception import InfrastructureException
+
+
+class DatabaseException(InfrastructureException):
+    """Error for database operation failures (RDS, connection issues, query exceptions)"""
