+++ norman_objects/shared/exceptions/internal_exceptions/infrastructure_exception.py
@@ -1,0 +1,5 @@
+from norman_objects.shared.exceptions.internal_exceptions.norman_internal_exception import NormanInternalException
+
+
+class InfrastructureException(NormanInternalException):
+    """Base class for infrastructure-related exceptions (AWS services, network, etc.)"""
